Consolidate modules in `events`.
