+++ backend/src/commands/processors/createGame.ts
@@ -1,8 +1,8 @@
 import * as cuid from 'cuid';
-import { saveEvent } from '../../events/store';
+import { saveEvent } from '../../events';
 import { createDeck, createTableau, shuffleDeck } from '../../game';
 import { GameCreatedEvent, GameEventType } from '../../events/types';
-import { CommandProcessor, CreateGameCommand} from '../types';
+import { CommandProcessor, CreateGameCommand } from '../types';
 
 export const createGame: CommandProcessor<CreateGameCommand, GameCreatedEvent> =
   async () => {
